Refactor: extract duplicated code
